Jesus tap dancing christ, the indentation in this library is a horror.
\t characters seem to be interpreted as 8 spaces, and the indent level ranges from 2-4 spaces, somewhat randomly, sometimes changing within a single file or even within a single function.  I don't know who wrote this thing, but they seriously need to learn how to operate whatever mickey mouse editor they use over there at the w3c.  *Every* editor nowadays supports set-and-forget indentation control.  Are they coding in Notepad? cat?  Srsly, WTF.  What kind of coder can stand to even look at this code, let alone write it and release it out into the world like this.

Seeing code like this, with such a low level of cleanliness, seriously makes me reconsider using it, and perhaps that's unfair.  But you know what?  Some people who don't floss are otherwise perfectly sanitary, but do you want a surgeon with crap in his teeth?  Didn't think so.

It's still full of visual fail all over the place, but at least with this global replacement, I won't have to keep changing the tabsize to 8, converting to spaces, and then changing back to 4.